+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.1.2
+
+* Added GROUP BY statement
+
+* Improved internal handling of WHERE clauses
+
 # 0.1.1
 
 * Fixed missing distinct in select queries